Small fix to Duck tide time-series
The tide timeseries are duck were slighly off as there was an artefact in the tidal model due to the think barrier island that is not fully resolved. I took a grid point further offshore to extract the tide and reprocessed the duck outputs for CoastSat and Shorex where tidal correction is applied.